Complete initial try at Rock Paper Scissors
I created an object to store the options. I feel like this allows for easy
expansion of the number of options at a later time. Each option has a list
of other options that it beats. This allows for easy input validation and win
condition validation.

I forgot to commit after creating the initial psuedocode so this is the first
attempt at writing the code using the written psuedocode but before testing.