Fix an error in checkWinCondition and in capitalizeString
The error in checkWinCondition was a logical error. I was trying to use 'in'
operator to check if a value was included in the array. This is valid in
Python but not here. Instead I need to use the array method 'includes'

The bug in capitalizeString was that I forgot to put the () at the end
of toUpperCase which caused the function itself to be displayed instead
of the results.