Add icons with a basic layout
I have the icons laid out on the page. The next step is adding the rest of the UI elements.

Afterwards I will setup the to be clickable and respond to the user clicking. I will need to assign an ID
to each of the images to pass to the javascript whenever the user clicks it.